+++ app/views/reviews/new.html.erb
@@ -31,12 +31,25 @@
           <h4><strong> Regarding the street... </strong></h4>
         </div>
 
-        <h4> <strong> " <%= @search.address %> </strong> " </h4>
-
+        <h4> <strong> " <%= @search.address %> "  </strong> </h4>
         <%= simple_form_for [@search, @review] do |f| %>
 
         <div style="padding-top: 30px;">
-          <h5> <%= f.input :street_review_average_rating, collection: 0..5, :label => "Your General Appreciation" %> </h5>
+          <h5>
+            <p>
+              <label style="margin-bottom: 25px;"><%= Answer::QUESTIONS[:q5][:title] %></label>
+            <input type="range" name="review[street_review_average_rating]" list="tickmarks" min="0" max="5" step="1">
+
+            <datalist id="tickmarks">
+              <option value="0" label="0">
+              <option value="1" label="2">
+              <option value="2" label="2">
+              <option value="3" label="3">
+              <option value="4" label="4">
+              <option value="5" label="5">
+            </datalist>
+            </p>
+           </h5>
           <h5> <%= f.input :street_review_title, :label => "Your Review Title" %> </h5>
           <h5> <%= f.input :street_review_content, :label => "Your Review" %> </h5>
           <h5> <%= f.input :photo, :label => "picture of your neighbourhood" %></h5>
@@ -49,12 +62,27 @@
       <div class="card-Review-Commune"">
         <div class="color-commune">
           <h4><strong>Regarding the commune...</strong></h4>
+
         </div>
 
         <h4> <strong> "<%= @commune.name %>" </strong> </h4>
 
         <div style="padding-top: 30px;">
-          <p> <%= f.input :commune_review_average_rating, collection: 0..5, :label => "Your General Appreciation" %> </p>
+          <p>
+            <p>
+              <label style="margin-bottom: 25px;"><%= Answer::QUESTIONS[:q7][:title] %></label>
+            <input type="range" name="review[commune_review_average_rating]" list="tickmarks" min="0" max="5" step="1">
+
+            <datalist id="tickmarks">
+              <option value="0" label="0">
+              <option value="1" label="2">
+              <option value="2" label="2">
+              <option value="3" label="3">
+              <option value="4" label="4">
+              <option value="5" label="5">
+            </datalist>
+            </p>
+           </p>
           <p> <%= f.input :commune_review_title, :label => "Your Review Title" %> </p>
           <p> <%= f.input :commune_review_content, :label => "Your Review" %> </p>
         </div>
