+++ app/views/shared/_mapradius.html.erb
@@ -91,4 +91,3 @@
     </div>
   </div>
   <% end %>
-
